feat(caesarCipher): Add cipher feature with tests
Tests for the caesarCipher test that it will loop
back to the start of the alphabet when the shift
factor would otherwise push it passed z. The tests
also handle symbols, spaces and retain upper and
lower case letters. Notably it'll also handle
shift factors greater than the alphabet length,
such that 'L' will become 'N'.