feat(analyzeArray): Add analyzeArray feat + tests
AnalyzeArray module will take an array of numbers
only and return an object containing the average
of the numbers, the minimum and maximum value.
Tests test for elements of the array to be numbers
as well as for the array to contain something.
They also test that each property we are after
is defined on the object being returned.